requirements.txt
celery==5.6.0
redis==7.1.0
django-celery-beat
django-celery-results

# OpenTelemetry Core Packages (Latest)
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0

# OpenTelemetry Instrumentations (Latest)
opentelemetry-instrumentation-django==0.46b0
opentelemetry-instrumentation-celery==0.46b0

gunicorn
whitenoise